+++ docs-src/app/App.svelte
@@ -1,13 +1,15 @@
 <Nav bind:component={component} />
+
+<ThemeSwitcher round class="docs-theme-switcher" />
+
 <main class={pageName}>
-	<Header />
 	<SvelteComponent/>
 </main>
 
 <script>
 import { Nav } from '../nav';
-import { Header } from '../header';
 import './App.css';
+import { ThemeSwitcher } from '../../src/theme-switcher';
 
 let component = $state();
 const pageName = $derived((component?.name ?? '').toLowerCase());