+++ src/input/button-toggle/types.ts
@@ -3,6 +3,7 @@ import type { InputProps } from '../types';
 export type ButtonToggleItem = {
 	name?: string,
 	value: string | boolean,
+	title?: string,
 	icon?: string;
 } | string;