+++ src/theme-switcher/index.ts
@@ -1,0 +1,3 @@
+export { default as ThemeSwitcher } from './ThemeSwitcher.svelte';
+export * from './types';
+export * from './utils';